requirements.txt
ImageIO==2.37.3
lazy-loader==0.5
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
scikit-image==0.26.0
scipy==1.17.1
tifffile==2026.3.3
